limitation tests

+++ test/snapshot-tests/createSchema.test.ts
@@ -45,10 +45,9 @@ function assertSchema(relativePath: string, options?: Options) {
             }schema.json`
         );
 
-        // if (!fs.existsSync(jsonFilePath)) {
-        fs.writeFileSync(jsonFilePath, JSON.stringify(schema, null, 4) + "\n", "utf8");
-        // }
-        // console.log(JSON.stringify(schema, null, 2));
+        if (!fs.existsSync(jsonFilePath)) {
+            fs.writeFileSync(jsonFilePath, JSON.stringify(schema, null, 4) + "\n", "utf8");
+        }
 
         const expected: any = JSON.parse(fs.readFileSync(jsonFilePath, "utf8"));
         const actual: any = JSON.parse(JSON.stringify(schema));
@@ -63,67 +62,57 @@ function assertSchema(relativePath: string, options?: Options) {
 }
 
 describe("createSchema", () => {
-    // it("react-class", assertSchema("react-class", { expose: "all" }));
-    // it("function-prop-type", assertSchema("function-prop-type"));
-    // it("module-function-declare", assertSchema("module-function-declare"));
-    // it("typescript-html-element-type", assertSchema("typescript-html-element-type"));
-    // it("extends-from-packages", assertSchema("extends-from-packages", { type: "MyProps" }));
-    // it("max-depth", assertSchema("max-depth", { maxDepth: 3 }));
-    // it("max-depth-ref", assertSchema("max-depth-ref", { maxDepth: 3 }));
-    // it(
-    //     "max-depth-function-param",
-    //     assertSchema("max-depth-function-param", { maxDepth: 3, type: "MyFunction", expose: "none" })
-    // );
-    // it(
-    //     "max-depth-function-param-ref",
-    //     assertSchema("max-depth-function-param-ref", { maxDepth: 3, type: "MyFunction" })
-    // );
-    // it("circular-ref-union", assertSchema("circular-ref-union", { type: "MyType", expose: "all" }));
-    // it("skipFiles option", assertSchema("skipFiles", { type: "MyType", skipParseTypeInFiles: ["external-props.ts"] }));
-    // it("skipTypes option", assertSchema("skipTypes", { type: "MyType", skipParseTypes: ["ExternalProps"] }));
-    // it(
-    //     "should parse type even if type is in skipFiles list",
-    //     assertSchema("skipFiles-with-forceToParseTypes", {
-    //         type: "MyType",
-    //         skipParseTypeInFiles: ["external-props.ts"],
-    //         forceToParseTypes: ["ExternalProps"],
-    //     })
-    // );
-    // it(
-    //     "should parse type even if type is in skipTypes list",
-    //     assertSchema("skipTypes-with-forceToParseTypes", {
-    //         type: "MyType",
-    //         skipParseTypes: ["ExternalProps"],
-    //         forceToParseTypes: ["ExternalProps"],
-    //     })
-    // // );
+    it("react-class", assertSchema("react-class", { expose: "all" }));
+    it("function-prop-type", assertSchema("function-prop-type"));
+    it("module-function-declare", assertSchema("module-function-declare"));
+    it("typescript-html-element-type", assertSchema("typescript-html-element-type"));
+    it("extends-from-packages", assertSchema("extends-from-packages", { type: "MyProps" }));
+    it("circular-ref-union", assertSchema("circular-ref-union", { type: "MyType", expose: "all" }));
+    it("skipFiles option", assertSchema("skipFiles", { type: "MyType", skipParseTypeInFiles: ["external-props.ts"] }));
+    it("skipTypes option", assertSchema("skipTypes", { type: "MyType", skipParseTypes: ["ExternalProps"] }));
+    it(
+        "should parse type even if type is in skipFiles list",
+        assertSchema("skipFiles-with-forceToParseTypes", {
+            type: "MyType",
+            skipParseTypeInFiles: ["external-props.ts"],
+            forceToParseTypes: ["ExternalProps"],
+        })
+    );
+    it(
+        "should parse type even if type is in skipTypes list",
+        assertSchema("skipTypes-with-forceToParseTypes", {
+            type: "MyType",
+            skipParseTypes: ["ExternalProps"],
+            forceToParseTypes: ["ExternalProps"],
+        })
+    );
 
     const dirs = getRelativeDirectories(resolve(`${basePath}/limit-options`));
-    // dirs.forEach(dir => {
-    //     it(
-    //         "excludeProperties" + dir,
-    //         assertSchema("limit-options/" + dir, {
-    //             type: "MyObject",
-    //             handleUnknownTypes: true,
-    //             excludeProperties: ["c", "a.b.c", "with-dash", "a.b.d.e", "x.d.e.f", "x.y", "a.b.x.c"],
-    //             expose: "none",
-    //             schemaExtension: "excludeProperties",
-    //         })
-    //     );
-    // });
+    dirs.forEach(dir => {
+        it(
+            "excludeProperties" + dir,
+            assertSchema("limit-options/" + dir, {
+                type: "MyObject",
+                handleUnknownTypes: true,
+                excludeProperties: ["c", "a.b.c", "with-dash", "a.b.d.e", "x.d.e.f", "x.y", "a.b.x.c"],
+                expose: "none",
+                schemaExtension: "excludeProperties",
+            })
+        );
+    });
 
-    // dirs.forEach(dir => {
-    //     it(
-    //         "includeProperties" + dir,
-    //         assertSchema("limit-options/" + dir, {
-    //             type: "MyObject",
-    //             handleUnknownTypes: true,
-    //             includeProperties: ["a.b", "c", "with-dash", "x"],
-    //             expose: "none",
-    //             schemaExtension: "includeProperties",
-    //         })
-    //     );
-    // });
+    dirs.forEach(dir => {
+        it(
+            "includeProperties" + dir,
+            assertSchema("limit-options/" + dir, {
+                type: "MyObject",
+                handleUnknownTypes: true,
+                includeProperties: ["a.b", "c", "with-dash", "x"],
+                expose: "none",
+                schemaExtension: "includeProperties",
+            })
+        );
+    });
 
     dirs.forEach(dir => {
         it(
@@ -138,13 +127,31 @@ describe("createSchema", () => {
         );
     });
 
-    // it(
-    //     "excludeProperties-interface",
-    //     assertSchema("limit-options", {
-    //         type: "MyObject",
-    //         handleUnknownTypes: true,
-    //         maxDepth: 1,
-    //         expose: "none",
-    //     })
-    // );
+    dirs.forEach(dir => {
+        it(
+            "includeProperties-maxDepth" + dir,
+            assertSchema("limit-options/" + dir, {
+                type: "MyObject",
+                handleUnknownTypes: true,
+                maxDepth: 2,
+                expose: "none",
+                schemaExtension: "includeProperties-maxDepth",
+                includeProperties: ["a"],
+            })
+        );
+    });
+
+    dirs.forEach(dir => {
+        it(
+            "excludeProperties-maxDepth" + dir,
+            assertSchema("limit-options/" + dir, {
+                type: "MyObject",
+                handleUnknownTypes: true,
+                maxDepth: 1,
+                expose: "none",
+                schemaExtension: "excludeProperties-maxDepth",
+                excludeProperties: ["c", "b"],
+            })
+        );
+    });
 });
